+++ src/common/hooks/useNavigationSettings/index.js
@@ -1,0 +1,5 @@
+export {
+  useNavigationSettings,
+  NAVIGATION_SETTINGS_SCOPE,
+  NAVIGATION_SETTINGS_BROWSE_TAB_KEY,
+} from './useNavigationSettings';